feat(timeSlider): extract responssive grid from panel  #1535
