+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
   "editor.detectIndentation": false,
-  "editor.renderWhitespace": "all",
+  "editor.renderWhitespace": "none",
   "editor.tabSize": 2,
   "editor.hover.delay": 1000,
   "window.title": "${activeEditorMedium}${separator}${rootName}${separator}${profileName}"
