+++ packages/geoview-core/src/core/components/common/index.ts
@@ -4,3 +4,4 @@ export * from './focus-trap-element';
 export * from './responsive-grid';
 export * from './layer-list';
 export * from './layer-title';
+export * from './layout';
